+++ CHANGES.md
@@ -27,6 +27,7 @@
 - Slowdown in `ensemble.predict_array` which caused the array to get sent to device in during each model evaluations
 -`Model.get_param_count` now includes mon-trainable params when requested
 - Fixed bug in `fold_lr_find` where LR finders would use different LR steps leading to NaNs when plotting in `fold_lr_find`
+- `plot_feat` used to coerce NaNs and Infs via `np.nan_to_num` prior to plotting, potentially impacting distributions, plotting scales, moments, etc. Fixed so that nan and inf values are removed rather than coerced.
 
 ## Changes